pyproject.toml
[tool.poetry]
authors = ["seria <seria.ati@gmail.com>"]
description = "PTT 爬蟲"
license = "GPL-3.0-only"
name = "ptt-scrape"
readme = "README.md"
version = "0.1.0"
package-mode = false

[tool.poetry.dependencies]
beautifulsoup4 = "^4.12.3"
lxml = "^5.3.0"
python = "^3.11"
requests = "^2.32.3"
loguru = "^0.7.2"
pydantic = "^2.8.2"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
ruff = "^0.5.7"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]
